Improve interface of RpcMessage types
Since Message and its wrapper types need to provide access to their held
inner value (ie Value object for Message types, and Message object for
the 3 Message wrapper types), the interface of message() and
raw_message() RpcMessage methods was both not clear, and not enough if
one wanted the inner Message type.

This patch changes the RpcMessage trait by renaming:

* message() -> as_vec()
* raw_message() -> as_value()

Additionally, a new trait RpcMessageType was added which is implemented
for RequestMessage, ResponseMessage, and NotificationMessage types. This
trait defines a method message() that provides a reference to the
Message wrapper's held Message object.