New type: Server
The Server type encapsulates a way to send messages/commands to a
running stream.

Currently, the type implements dual channels, one for the purpose of
sending and receiving server messages, and a second channel for yielding
data from the stream. The basic idea is that incoming connections would
yield sockets all the while listening for any incoming commands.

Currently, only messages for yielding sockets from the stream, and for
shutting down the stream are implemented.