FIX: storage/lmdb: mode type on unix and windows is not the same
The issue was hardcoding a single type for representing the db file
mode. However, this type is different depending on the platform being
run.

This patch fixes the issue by setting the type to a newtype `mode_t`
that's defined by the `lmdb-sys` crate. This ensures that the correct
type is always used for the platform the program is compiled to.