FIX: server shutting down after session completes
The expected behaviour is to only close the current connection when a
session completes, or some error occurs (eg invalid message) and leave
the server running.

This fix retains the ability of the message codec and the state machine
to shutdown the server, but now will only close the connection during
normal operation.